Move the DatePickerComponent to the components folder
